style: Conform to the styling guides (GNU 89 and Betty)
