+++ main.h
@@ -22,6 +22,9 @@ char *p_concat(char *path, char *command);
 char  *s_concat(char *str1, char *str2);
 int _strlen(char *str);
 char *resolve_r_path(char *input);
-const char* get_env_value(const char* name);
+const char *get_env_value(const char *name);
+
+
+extern char **environ;
 
 #endif
