+++ misc_func.c
@@ -1,5 +1,11 @@
 #include "main.h"
 
+/**
+ * path - A function that gets the path to a command
+ * @input: The input that is gotten from stdin
+ * Return: The absolute path to the command
+*/
+
 char *path(char *input)
 {
 	char *cmd = strdup(input);
@@ -11,18 +17,17 @@ char *path(char *input)
 	struct stat file_exist;
 	char bash_variables[] = {'.', '/', '~'};
 	size_t i = 0;
-	char r_path[PATH_MAX];
 
 	char *p_part = strtok(path_arr_cpy, delim);
 
 	while (p_part)
 	{
 		char *fp_cmd = p_concat(p_part, cmd);
+
 		if (stat(fp_cmd, &file_exist) == 0)
 		{
 			t_path = fp_cmd;
 			return (t_path);
-			break;
 		}
 
 		p_part = strtok(NULL, delim);
@@ -40,6 +45,13 @@ char *path(char *input)
 
 	return (t_path);
 }
+
+/**
+ * resolve_r_path - Gets relative path and converts it into
+ * absolute path
+ * @input: The command given by user
+ * Return: The absolute path of the command
+*/
 
 char *resolve_r_path(char *input)
 {
@@ -64,77 +76,4 @@ char *resolve_r_path(char *input)
 	realpath(input, fr_path);
 
 	return (fr_path);
-}
-
-
-char *p_concat(char *path, char *command)
-{
-    char *new_str = malloc(sizeof(char) * (_strlen(path) + _strlen(command) + 1));
-    int i = 0, j = 0;
-    while (path[i])
-    {
-        new_str[i] = path[i];
-        i++;
-    }
-
-    new_str[i] += '/';
-    i++;
-
-    while (command[j])
-    {
-        new_str[i] += command[j];
-        i++;
-        j++;
-    }
-
-    new_str[i] = '\0';
-
-    return (new_str);
-}
-
-const char* get_env_value(const char* name) {
-     size_t name_len = strlen(name);
-	 extern char **environ;
-     char** env_ptr = environ;
-
-     while (*env_ptr) {
-         if (strncmp(*env_ptr, name, name_len) == 0 && (*env_ptr)[name_len] == '=')
-             return *env_ptr + name_len + 1;
-         env_ptr++;
-     }
-
-     return NULL;
-}
-
-
-int _strlen(char *str)
-{
-	size_t i = 0;
-
-	for (i = 0; *str; i++)
-		str++;
-
-	return (i);
-}
-
-char  *s_concat(char *str1, char *str2)
-{
-    char *new_str = malloc(sizeof(char) * (_strlen(str1) + _strlen(str2) + 1));
-    int i = 0, j = 0;
-    while (str1[i])
-    {
-        new_str[i] = str1[i];
-        i++;
-    }
-
-    while (str2[j])
-    {
-        new_str[i] += str2[j];
-        i++;
-        j++;
-    }
-
-    new_str[i] = '\0';
-
-    return (new_str);
 }
